+++ tests/manual/wulff_net.py
@@ -1,7 +1,7 @@
 import numpy as np
 import matplotlib.pyplot as plt
 
-from ipf import cubic
+from crystalipf import cubic
 
 
 if __name__ == "__main__":
